Separate MAX7219 matrix from the general driver
